all: use GoReleaser for releases
