+++ Makefile
@@ -13,10 +13,6 @@ export GOPKG=$(PROJECTPATH)/pkg
 GO_LDFLAGS=""
 DIST=$(PROJECTPATH)/dist
 
-ifeq ($(GOOS),windows)
-BIN_EXT = .exe
-endif
-
 project = heminetwork
 version = $(shell git describe --tags 2>/dev/null || echo "v0.0.0")
 
@@ -29,16 +25,13 @@ cmds = \
 	popmd	\
 	tbcd
 
-.PHONY: all clean clean-dist deps go-deps $(cmds) build install lint lint-deps tidy race test vulncheck \
-	vulncheck-deps dist archive sources checksums
+.PHONY: all clean deps go-deps $(cmds) build install lint lint-deps tidy race test vulncheck \
+	vulncheck-deps
 
 all: lint tidy test build install
 
-clean: clean-dist clean-test
+clean: clean-test
 	rm -rf $(GOBIN) $(GOCACHE) $(GOPKG)
-
-clean-dist:
-	rm -rf $(DIST)
 
 clean-test:
 	rm -rf $(PROJECTPATH)/service/tbc/.testleveldb/
@@ -90,19 +83,3 @@ vulncheck:
 
 vulncheck-deps:
 	GOBIN=$(shell go env GOPATH)/bin go install golang.org/x/vuln/cmd/govulncheck@latest
-
-dist:
-	mkdir -p $(DIST)
-
-archive: dist install
-ifeq ($(GOOS),windows)
-	cd $(GOBIN) && zip -r $(DIST)/$(project)_$(version)_$(GOOS)_$(GOARCH).zip *$(BIN_EXT)
-else
-	cd $(GOBIN) && tar -czvf $(DIST)/$(project)_$(version)_$(GOOS)_$(GOARCH).tar.gz *
-endif
-
-sources: dist
-	tar --exclude=dist --exclude=bin -czvf $(DIST)/$(project)_$(version)_sources.tar.gz * .gitignore .github
-
-checksums: dist
-	cd $(DIST) && shasum -a 256 * > $(project)_$(version)_checksums.txt
